add interactions with mob
add zenject
add part of house func

+++ Assets/Scripts/Configs/Mobs/BrainrotMobConfig.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class BrainrotMobConfig : ScriptableObject
 {
     [field: SerializeField] public string Name { get; private set; }
-    [field: SerializeField] public float BaseCost { get; private set; }
+    [field: SerializeField] public int BaseCost { get; private set; }
     [field: SerializeField] public float ValuePerSecond { get; private set; }
     [field: SerializeField] public GameObject MobPrefab { get; private set; }
 }
